+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this extension pack will be documented in this file.
 
 ## 0.0.1
 
+- Added Automatic run/debug execution [#5](https://github.com/rodrigotol/vscode-go-pack-go/pull/5)
 - Added Run/debug delegation [#4](https://github.com/rodrigotol/vscode-go-pack-go/pull/4)
 - Added CodeLens integration [#3](https://github.com/rodrigotol/vscode-go-pack-go/pull/3)
 - Added Backward detection algorithm [#2](https://github.com/rodrigotol/vscode-go-pack-go/pull/2)